requirements.txt
appdirs==1.4.4
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudscraper==1.2.71
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
curl_cffi==0.7.4
dataclasses==0.6
defusedxml==0.7.1
dnspython==2.7.0
EbookLib==0.18
filelock==3.16.1
filetype==1.2.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parsel==1.9.1
pillow==11.0.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.1
pymongo==4.10.1
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyppeteer==2.0.0
pypub3==2.0.7
python-dateutil==2.9.0.post0
pytz==2024.2
pyxml3==0.0.4
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
Scrapy==2.11.2
scrapy-splash==0.11.1
scrapy-user-agents==0.1.1
service-identity==24.1.0
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
tldextract==5.1.2
tqdm==4.67.1
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==1.26.20
user-agents==2.2.0
w3lib==2.2.1
websockets==10.4
zipp==3.21.0
zope.interface==7.1.1
